Move TimeSeriesData from opt/ to utils/ to fix inverted dependency
TimeSeriesData is a general data container used by both blocks/ and opt/.
Having blocks/ import from opt/ violated the expected layering.

- Move canonical definition to src/pathsim/utils/timeseries_data.py
- Delete opt/timeseries_data.py
- Update all import sites: opt/__init__.py, opt/parameter_estimator.py,
  blocks/timeseries_source.py, and the two test files
