+++ src/pathsim/utils/timeseries_data.py
